abstracted logic for email api
